split `MainLayout`
